refactor(components): correct header deprecation notice and overflow claims
