+++ .changeset/scalar-header-chrome-only.md
@@ -2,9 +2,9 @@
 '@scalar/components': minor
 ---
 
-**Breaking:** `ScalarHeader` no longer lays out its own columns. The `start`, `center` and `end` slots have been removed in favour of composing `ScalarHeaderColumn` children, and a new `is` prop lets the header render as something other than a `header` element.
+**Breaking:** `ScalarHeader` no longer lays out its own columns. The `start` and `end` slots have been removed in favour of composing `ScalarHeaderColumn` children in the default slot, and a new `is` prop lets the header render as something other than a `header` element.
 
-Columns hug their content. Give a column `flex-1` to let it take the free space, and give both side columns `flex-1` when the middle one should sit centred.
+Columns hug their content. Give the content-bearing column `flex-1` so it takes the free space — that is also the column that absorbs the shrinking when space runs out.
 
 ```diff
   <ScalarHeader>
@@ -23,6 +23,18 @@ Columns hug their content. Give a column `flex-1` to let it take the free space,
   </ScalarHeader>
 ```
 
-Passing the old slots logs a deprecation warning, and typed consumers get a `vue-tsc` error. One case cannot be detected: a header that used *only* the centre slot now renders its contents left-aligned rather than centred, because that content is indistinguishable from the new default slot. Wrap it in a `ScalarHeaderColumn` with `justify-center` and `flex-1` on the columns either side.
+**Put `flex-1` on both side columns only when there is a genuine `justify-center` middle column.** In a two-column header it splits the width 50/50 instead, which clips the content-bearing side on narrow screens:
 
-This also fixes the header overflowing narrow viewports, including on iOS Safari. `ScalarHeaderColumn` carries `min-w-0`, so a long title ellipsizes instead of pushing the header past the screen.
+```html
+<ScalarHeader>
+  <ScalarHeaderColumn class="flex-1">…</ScalarHeaderColumn>
+  <ScalarHeaderColumn class="justify-center">…</ScalarHeaderColumn>
+  <ScalarHeaderColumn class="flex-1 justify-end">…</ScalarHeaderColumn>
+</ScalarHeader>
+```
+
+Passing the old slots logs a deprecation warning, and typed consumers get a `vue-tsc` error. One case cannot be detected: a header that used *only* the middle slot now renders its contents left-aligned rather than centred, because that content is indistinguishable from the new default slot. Wrap it as above.
+
+`is` is worth reaching for when a header is nested inside an existing `header` landmark — `is="div"` avoids exposing a second `banner` role.
+
+This also fixes the app header overflowing narrow viewports, including on iOS Safari. `ScalarHeaderColumn` carries `min-w-0`, so a `flex-1` column shrinks and its long title ellipsizes instead of pushing the header past the screen. Hug columns still size to their content, so keep a trailing action cluster narrow enough to fit on its own.
