+++ packages/components/src/components/ScalarHeader/ScalarHeader.vue
@@ -3,7 +3,12 @@
  * Scalar header component
  *
  * Provides the header chrome; layout is left to the consumer. Columns hug their
- * content, so give one `flex-1` to take the free space.
+ * content, so give the content-bearing one `flex-1` to take the free space - that
+ * is also the column that absorbs the shrinking when space runs out.
+ *
+ * Headers that used the old middle slot are no longer centred: wrap that content
+ * in a `ScalarHeaderColumn` with `justify-center` and give both side columns
+ * `flex-1`.
  *
  * @example
  * ```html
@@ -18,6 +23,9 @@
  * ```
  */
 export default {}
+
+/** Deprecation notices are about the API, not the instance, so warn once */
+const hasWarned = false
 </script>
 <script setup lang="ts">
 import { useBindCx } from '@scalar/use-hooks/useBindCx'
@@ -38,10 +46,12 @@ const { cx } = useBindCx()
 const slots = useSlots()
 
 // Vue drops content for slots we no longer render, so an unmigrated consumer
-// would otherwise just get an empty header.
-if (slots.start || slots.end) {
+// would otherwise just get an empty header. Warn once per module rather than
+// per instance - a header can mount many times over a session.
+if (!hasWarned && (slots.start || slots.end)) {
+  hasWarned = true
   console.warn(
-    'ScalarHeader: the `start`, `center` and `end` slots have been removed. Compose `ScalarHeaderColumn` children instead.',
+    'ScalarHeader: the `start` and `end` slots have been removed. Compose `ScalarHeaderColumn` children in the default slot instead.',
   )
 }
 </script>
