+++ packages/components/src/components/ScalarHeader/ScalarHeaderColumn.vue
@@ -5,8 +5,9 @@
  * A section of a `ScalarHeader`. Columns hug their content, so pass `flex-1` to
  * take the free space and `justify-center` / `justify-end` to place the contents.
  *
- * Carries `min-w-0` so a column can shrink below its contents and long titles
- * ellipsize instead of widening the header.
+ * Carries `min-w-0`, which lets a `flex-1` column shrink below its contents so a
+ * long title ellipsizes rather than widening the header. Hug columns do not
+ * shrink, so keep a trailing action cluster narrow enough to fit on its own.
  *
  * @example
  * ```html
